# Introduce speculative decoding with draft models to vLLM