--- a/requirements.txt
+++ b/requirements.txt
@@ -13,3 +13,4 @@ aioprometheus[starlette]
 pynvml == 11.5.0
 triton >= 2.1.0
 cupy-cuda12x == 12.1.0  # Required for CUDA graphs. CUDA 11.8 users should install cupy-cuda11x instead.
+flash-attn >= 2.5.0