--- a/vllm/config.py
+++ b/vllm/config.py
@@ -447,6 +447,7 @@ class SchedulerConfig:
         max_model_len: Maximum length of a sequence (including prompt
             and generated text).
         max_paddings: Maximum number of paddings to be added to a batch.
+        parallel_decoding_lookahead: Number of tokens to look ahead for parallel decoding.
     """
 
     def __init__(
@@ -455,6 +456,7 @@ class SchedulerConfig:
         max_num_seqs: int,
         max_model_len: int,
         max_paddings: int,
+        parallel_decoding_lookahead: Optional[int] = 1,
     ) -> None:
         if max_num_batched_tokens is not None:
             self.max_num_batched_tokens = max_num_batched_tokens
@@ -465,6 +467,7 @@ class SchedulerConfig:
         self.max_num_seqs = max_num_seqs
         self.max_model_len = max_model_len
         self.max_paddings = max_paddings
+        self.parallel_decoding_lookahead = parallel_decoding_lookahead
         self._verify_args()
 
     def _verify_args(self) -> None: