This PR is (yet another) implementation of speculative decoding on vLLM, compared with existing efforts (including #2607, #1797, and #1679), this PR:

- completed, confirmed workable (and correct)
- might be the most up-to-date implementation (based on current main)
- simplest
  - without adding new kernels (`context_attention_fwd` from the prefix cache PR is enough, add is compatible with further effort of introducing flash-attn and flashinfer)
  - reusing existing components: the rejected sampler (from #2336) and multi-step worker (from #2424), without reinvent a new wheel (thanks you! @cadedaniel)
- support (paged) kv-cache for draft model as well
- compatible with existing features, including paged kv-cache and the prefix cache feature

(Note that this PR is built-upon the PR #3007 (GQA fixes for `context_attention_fwd`) and #3010 (introducing flash-attn to vLLM), but not depends on these two PRs. If these two PRs can be accepted, I would rebase this PR then, otherwise submitting the speculative sampling feature in a separate PR also works for me.)

The major change only happens in `llm_engine.py`'s `step()` method, `model_runner.py`'s `_prepare_decode()/_prepare_sample()` method, should be fairly easy for code review.

**The major design & implementation can be highlighted as follows**:

- KV cache:
  - for target model and draft model: the usable memory for KV cache is splitted by respecting the per-token kv-cache size's proportion of the target and draft model.
    - i.e., the `num_gpu_blocks` is computed from `gpu_memory / (draft_block_size + target_block_size)`
  - The draft model and target block shares the same block table and slot mapping, simplifying implementation in scheduler
- The scheduler is responsible for choosing sequences to run (as usually), and the `LLMEngine`'s `step()` method will run draft model for k times and then a target model step follows.
- Attention:
  - Specualtive decoding requires verifying k tokens at the same time, this is implemented using the `context_attention_fwd` kernel (originally added for prefix caching)
  - Compatible with flash-attn's `flash_attn_with_kvcache` kernel.
- Sampling:
  - This PR follows the same logic with the `transformer` package for how to decide if a token should be accepted, more specifically,
    - using reject sampler when do random sampling
    - comparing if tokens are the same when do greedy sampling (i.e., `temperature=0.0`)
    - see also: https://github.com/huggingface/transformers/blob/45244940725ec1b3e4c390b74dbafe65b298acca/src/transformers/generation/utils.py#L4566-L4597

TODO:

- The interaction of speculative decoding and beam search is a bit complicate (see also the original ICML paper's appendix 4). This PR raising `NotImplementedError` for such case and leave it as a TODO.
- ~~Support initialize draft workers on Ray.~~

Numbers of a prompt randomly choosed from dataset, using Llama-2-70B-GPTQ as the target model and Llama-2-7B-GPTQ/TinyLLama-1.1B-Chat-v1.0-GPTQ as the draft model:

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:////Users/linzhu.ht/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip.htm">
<link rel=File-List
href="file:////Users/linzhu.ht/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip_filelist.xml">
</head>

<body link="#0563C1" vlink="#954F72">


Model | DraftModel | Quantization | Prompt | Generation | BatchSize | Lookahead | Drafted | Accepted | AcceptRate | TG Speed (tokens/second) | TG Latency (seconds)
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
70B | 1.1B | GPTQ | 10 | 200 | 1 | 1 | 0 | 0 | 0 | 33.8 | 0.0295858
70B | 7B | GPTQ | 10 | 200 | 1 | 1 | 0 | 0 | 0 | 33.9 | 0.02949853
70B | 1.1B | GPTQ | 10 | 200 | 1 | 2 | 109 | 85 | 77 | 43.3 | 0.02309469
70B | 7B | GPTQ | 10 | 200 | 1 | 2 | 107 | 93 | 86 | 42.8 | 0.02336449
70B | 1.1B | GPTQ | 10 | 200 | 1 | 3 | 162 | 113 | 69 | 53.1 | 0.01883239
70B | 7B | GPTQ | 10 | 200 | 1 | 3 | 149 | 121 | 81 | 51.4 | 0.01945525
70B | 1.1B | GPTQ | 10 | 200 | 1 | 4 | 197 | 116 | 58 | 53.6 | 0.01865672
70B | 7B | GPTQ | 10 | 200 | 1 | 4 | 203 | 150 | 73 | 47.8 | 0.0209205
70B | 1.1B | GPTQ | 10 | 200 | 1 | 5 | 234 | 136 | 58 | 52.3 | 0.01912046
70B | 7B | GPTQ | 10 | 200 | 1 | 5 | 205 | 166 | 80 | 52.6 | 0.01901141
70B | 1.1B | GPTQ | 10 | 200 | 1 | 6 | 262 | 153 | 58 | 47 | 0.0212766
70B | 7B | GPTQ | 10 | 200 | 1 | 6 | 260 | 180 | 69 | 42 | 0.02380952
70B | 1.1B | GPTQ | 10 | 200 | 1 | 7 | 295 | 154 | 52 | 46.2 | 0.02164502
70B | 7B | GPTQ | 10 | 200 | 1 | 7 | 269 | 173 | 64 | 45.7 | 0.02188184
70B | 1.1B | GPTQ | 10 | 200 | 1 | 8 | 707 | 268 | 38 | 39.7 | 0.02518892
70B | 7B | GPTQ | 10 | 200 | 1 | 8 | 274 | 184 | 67 | 46.9 | 0.02132196
70B | 1.1B | GPTQ | 10 | 200 | 64 | 1 | 0 | 0 | 0 | 283 | 0.22614841
70B | 7B | GPTQ | 10 | 200 | 64 | 1 | 0 | 0 | 0 | 283 | 0.22614841
70B | 1.1B | GPTQ | 10 | 200 | 64 | 2 | 31872 | 24621 | 78 | 427.7 | 0.1496376
70B | 7B | GPTQ | 10 | 200 | 64 | 2 | 29632 | 26226 | 90 | 422.9 | 0.15133601
70B | 1.1B | GPTQ | 10 | 200 | 64 | 3 | 41088 | 29947 | 76 | 578.1 | 0.11070749
70B | 7B | GPTQ | 10 | 200 | 64 | 3 | 36608 | 29547 | 83 | 490.5 | 0.1304791
70B | 1.1B | GPTQ | 10 | 200 | 64 | 4 | 41344 | 25872 | 63 | 610.7 | 0.10479777
70B | 7B | GPTQ | 10 | 200 | 64 | 4 | 36544 | 26938 | 76 | 562.2 | 0.11383849
70B | 1.1B | GPTQ | 10 | 200 | 64 | 5 | 45760 | 27635 | 62 | 640.7 | 0.09989074
70B | 7B | GPTQ | 10 | 200 | 64 | 5 | 40704 | 30531 | 76 | 548.3 | 0.11672442
70B | 1.1B | GPTQ | 10 | 200 | 64 | 6 | 55872 | 28106 | 51 | 661.3 | 0.09677907
70B | 7B | GPTQ | 10 | 200 | 64 | 6 | 47040 | 31387 | 69 | 597 | 0.10720268
70B | 1.1B | GPTQ | 10 | 200 | 64 | 7 | 59520 | 26615 | 50 | 654.3 | 0.09781446
70B | 7B | GPTQ | 10 | 200 | 64 | 7 | 66560 | 40516 | 64 | 612 | 0.10457516
70B | 1.1B | GPTQ | 10 | 200 | 64 | 8 | 65408 | 26316 | 44 | 710.4 | 0.09009009
70B | 7B | GPTQ | 10 | 200 | 64 | 8 | 71872 | 41465 | 60 | 593.1 | 0.1079076



</body>

</html>