@nivibilla 2024-02-25
Hi, this doesn't seem to work with ray atm. I was trying to benchmark Mixtral8x7b with Mistral 7b as the draft model.

`AssertionError: Speculative decoding is not supported with Ray.`