@nivibilla 2024-02-25
also im getting this error when trying to run

`ModuleNotFoundError: No module named 'vllm.worker.spec_decode'`

Im installing like

`!pip install git+https://github.com/sighingnow/vllm.git@ht/speculative`