@sighingnow 2024-02-25
> `ModuleNotFoundError: No module named 'vllm.worker.spec_decode'`

Should be fixed now. Ray worker initialization has been added as well (not tested on ray, but it basically shares the same logic as before).