@sighingnow 2024-02-25
> Just tried this again, with tp 8. It freezes after

Thanks for the feedback. Could you please help to confirm that without specualtive decoding (and the draft model), the model `Mixtral-8x7B-Instruct-v0.1-HF` can work as expected? Thanks!